Commit made by the Bioconductor Git-SVN bridge. Consists of 1 commit.
Commit information:

Commit id: e2d415b18736ff2179381b5005bbe9efc9810155
Commit message: change, where appropriate, to size_t arguments in c code
Committed by: R building
Author Name: R building
Commit date: 2014-08-31 15:25:49 -0700
Author date: 2014-08-31 15:25:49 -0700



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/preprocessCore@93721 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/avg.h
@@ -1,10 +1,10 @@
 #ifndef AVG_H
 #define AVG_H
 
-void ColAverage(double *data, int rows, int cols, int *cur_rows, double *results, int nprobes, double *resultsSE);
-void ColAverage_noSE(double *data, int rows, int cols, int *cur_rows, double *results, int nprobes);
+void ColAverage(double *data, size_t rows, size_t cols, int *cur_rows, double *results, size_t nprobes, double *resultsSE);
+void ColAverage_noSE(double *data, size_t rows, size_t cols, int *cur_rows, double *results, size_t nprobes);
 
-void colaverage_no_copy(double *data, int rows, int cols, double *results, double *resultsSE);
-void colaverage(double *data, int rows, int cols, double *results, double *resultsSE);
+void colaverage_no_copy(double *data, size_t rows, size_t cols, double *results, double *resultsSE);
+void colaverage(double *data, size_t rows, size_t cols, double *results, double *resultsSE);
 
 #endif
